+++ zoo/src/main/scala/com/intel/analytics/zoo/examples/objectdetection/finetune/ssd/Train.scala
@@ -24,9 +24,9 @@ import com.intel.analytics.bigdl.utils.LoggerFilter
 import com.intel.analytics.bigdl.visualization.{TrainSummary, ValidationSummary}
 import com.intel.analytics.zoo.common.NNContext
 import com.intel.analytics.bigdl.nn.Module
+import com.intel.analytics.zoo.models.image.objectdetection.common.dataset.SSDMiniBatch
 import com.intel.analytics.zoo.models.image.objectdetection.common.{IOUtils, MeanAveragePrecision, ModuleUtil}
 import com.intel.analytics.zoo.models.image.objectdetection.common.nn.MultiBoxLoss
-import com.intel.analytics.zoo.models.image.objectdetection.common.dataset.roiimage.SSDMiniBatch
 import com.intel.analytics.zoo.models.image.objectdetection.common.nn.MultiBoxLossParam
 import com.intel.analytics.zoo.models.image.objectdetection.ssd.SSDVgg
 import org.apache.log4j.{Level, Logger}
