+++ zoo/src/main/scala/com/intel/analytics/zoo/models/image/objectdetection/common/dataset/RoiByteImageToSeq.scala
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.zoo.models.image.objectdetection.common.dataset.roiimage
+package com.intel.analytics.zoo.models.image.objectdetection.common.dataset
 
 import java.io.File
 import java.nio.ByteBuffer
 import java.nio.file.Path
 
-import com.intel.analytics.bigdl.dataset.Transformer
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.transform.vision.image.ImageFeature
 import com.intel.analytics.bigdl.transform.vision.image.label.roi.RoiLabel
 import com.intel.analytics.zoo.feature.common.Preprocessing
-import com.intel.analytics.zoo.models.image.objectdetection.common.dataset.roiimage.RoiByteImageToSeq._
 import org.apache.commons.io.FileUtils
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{Path => hPath}
