+++ zoo/src/main/scala/com/intel/analytics/zoo/models/image/objectdetection/common/dataset/RoiImageSeqGenerator.scala
@@ -20,7 +20,6 @@ import java.io.File
 import java.nio.file.Paths
 
 import com.intel.analytics.zoo.models.image.objectdetection.common.IOUtils
-import com.intel.analytics.zoo.models.image.objectdetection.common.dataset.roiimage.RoiByteImageToSeq
 import scopt.OptionParser
 
 object RoiImageSeqGenerator {
